[add]バックエンドのディレクトリ構成整理とDB接続実装
