+++ backend/Dockerfile
@@ -1,8 +1,10 @@
 FROM golang:1.16.3
 
-WORKDIR /go/src
-COPY ./src /go/src
+WORKDIR /app
+COPY . /app
 
-RUN go get github.com/labstack/echo/
+RUN go get github.com/labstack/echo
+RUN go get gorm.io/driver/mysql
+RUN go get gorm.io/gorm
 
-CMD ["go", "run", "server.go"]
+CMD ["go", "run", "main.go"]
