--- backend/src/go.mod
@@ -1,5 +1,0 @@
-module github.com/SyadoWCS/single_tournament
-
-go 1.16
-
-require github.com/labstack/echo/v4 v4.4.0
