--- backend/src/server.go
@@ -1,19 +1,0 @@
-package main
-
-import (
-    "net/http"
-    "github.com/labstack/echo/v4"
-)
-
-func main() {
-    // Echoのインスタンス作る
-    e := echo.New()
-
-    // ルーティング
-    e.GET("/", func(c echo.Context) error {
-        return c.String(http.StatusOK, "Hello World")
-    })
-
-    // サーバー起動
-    e.Start(":80")
-}
