+++ docker-compose.yml
@@ -7,6 +7,11 @@ services:
       context: ./backend
     ports:
       - 80:80
+    environment:
+      MYSQL_USER: ${MYSQL_USER}
+      MYSQL_PASSWORD: ${MYSQL_PASSWORD}
+      MYSQL_ENDPOINT: ${MYSQL_ENDPOINT}
+      MYSQL_DATABASE: ${MYSQL_DATABASE}
 
   frontend:
     container_name: single-tournament-frontend
